+++ src/tree.c
@@ -49,12 +49,12 @@ void release_huffman_tree(struct tree_node* root) {
 
 
 
-void count_huffman_codes(struct tree_node* root, uint32_t huff_code[], uint8_t huff_code_length[]) {
+void count_huffman_codes(struct tree_node* root, uint32_t huff_code[], uint32_t huff_code_length[]) {
      _count_huffman_codes_helper(root, 0, 0, huff_code, huff_code_length);
 }
 
-void _count_huffman_codes_helper(struct tree_node* elem, uint8_t level, uint32_t code,
-                                 uint32_t huff_code[], uint8_t huff_code_length[]) {
+void _count_huffman_codes_helper(struct tree_node* elem, uint32_t level, uint32_t code,
+                                 uint32_t huff_code[], uint32_t huff_code_length[]) {
      if (elem->left_son == NULL && elem->right_son == NULL) {
           huff_code[elem->value] = code;
           huff_code_length[elem->value] = level;
